jwt 리팩토링

+++ src/main/java/org/project/study/config/JwtSettings.java
@@ -14,5 +14,5 @@ public class JwtSettings {
 
     private String tokenSigningKey;
 
-    private Integer refreshTokenExpTime;
+    private Long refreshTokenExpTime;
 }
